+++ lib/rules.dart
@@ -5,3 +5,4 @@ export 'package:cellular_automata/src/rules/_ca_rules.dart';
 export 'package:cellular_automata/src/rules/game_of_life.dart';
 export 'package:cellular_automata/src/rules/game_of_life_simple.dart';
 export 'package:cellular_automata/src/rules/brians_brain.dart';
+export 'package:cellular_automata/src/rules/majority_vote.dart';